requirements.txt
numpy
scipy
Pillow
wxPython
wxPython-common
ObjectListView 
traits
traitsui
mayavi
PyYAML
